+++ changelog.md
@@ -3,6 +3,9 @@
 #### V0.1.1
  + Show short definitions above synonyms
  + Automatically adjust popup positioning after content is added so that it doesn't overflow off side or bottom of page
+ + Replace text with selected synonym on textarea and input elements
+ + "enable only on editable text" setting working
+ + Small improvements
 
 #### V0.1.0
  + Basic synonym popup with a list of synonyms and a close button